Update schema and README to reflect new process for creating the db
This should resolve the issue of the collation / locale issue that
appears to be different on different operating systems. The locale
should get set appropriately when the database is created and so the
script no longer needs to handle that.

The README now includes instructions on how to set up a separate user
for the database which is more appropriate for production.