Load database config in app.js
It may need to be used by code outside of the "routes" hierarchy, such
as the socket.io code that is listening for new lines coming in? Plus
this way nothing happens if the database config is not found and it's
obvious to the reader that this config is required.