User registration (no password hashing yet)
Call /register POST from the registration page. Insert the username,
email, and (unhashed) password into the poets table.

Since passwords aren't yet hashed and the column is a char(64), you have
to pad the password with trailing spaces to make it exactly 64 characters
if you want to log in at this point. It won't matter, because password hashing
coming up next. Briefly considered trimming the password, but since we
don't want to be storing passwords in the clear anyway, we probably won't
need this logic.