Hash passwords on register and login.
SHA256 hash of all passwords is performed client-side before registration
and login. jQuery is used to intercept form submission and write the
hashed password to a hidden field.

There is simple validation in place to ensure that the password matches
the confirmed password, but no nice error message yet. Validation is
basically non-existent at this point. Users can submit blank passwords,
and there is no server-side validation that the password is actually
SHA256.

A user is placed on the request by Passport, and this is passed through
to the poem list page. The userId does not need to appear in the url,
since it can be accessed via the request object. No globals necessary,
either.