Save lines to database
Move the socket.io poem editor into its own module. Hook it up to the
linesRepository, and clean up the event names. line-submitted is triggered
by the client when a new line is written, and line-created is broadcast by
the server when the line is saved to the database.

Move the HTML escaping logic to the client, so the raw data can be stored
in the database. The saved lines should not be coupled to any specific
escaping logic, so it makes sense to escape them as late as possible (on
the client). The jade template will also escape any tags.

The poetId and poemId are encoded in data attributes on the form tag. The
jQuery logic reads the data attributes, then triggers the line-submitted
event. In the line-created handler, a more modern jQuery syntax is used to
create the .line divs.