Improved application error handling
Errors are currently classified as one of 3 types:

1) Handled Errors: in the routes/common respond method, these are "nice"
errors that happen when the respond method is given a value in the err
parameter. We should strive to make most of our errors come through this
block.

2) Unhandled Exceptions: the routes/common respond method catches
exceptions that happen in the success callback.

3) Unhandled Errors: these are errors that happen in middleware components.
errors in jade templates, for example, will result in errors that are
passed through to the error handling middleware function in app.js.

All of these errors are handled in the same way: an HTTP 500 is sent to
the client, and a stack trace is logged on the server. No error details
are delivered to the client, to avoid internal implementation disclosure.