Poet typeahead and poet search
Add search method to poets repository, and expose it at the /poet?q
endpoint. It finds the first 10 poets that start with the specified
name.

Add a Poet search input to the new poem form and hook it up to the
search using the twitter typeahead library. Poets are added to a ul
but are not yet associated with the poem when it is saved.