Save poet to poem associations
Create a hidden input with the poet id when a poet is added to the
typeahead. An array of poet ids is submitted to the server.

New addPoets method on the poemsRepository accepts an array of poet_ids
and uses postgres' spiffy ANY operator to insert all poet_ids in a single
SQL statement.