Exclude poets from typeahead that have already been added to the poem.
Use the typeahead filter function to exclude items that are already in the list. This approach has the limitation of potentially returning fewer than 10 results. In a worst-case scenario, the top 10 poets starting with a given letter could be added to the poem, and 0 results would be returned! Another approach could be to pass the list of poets to the server and exclude them with a SQL query. This would probably be prohibatively expensive from a performance standpoint.