Add the user to the new poem by default.
Render the user in the poet list on the new poem page. Remove the server-side logic that adds the current user to the poem, as this is effectively done on the client now. There is now some conceptual duplication between the server and the client, because the jade template and the jquery logic both need to know the structure of the li and the containing input.

Also changed the wording of the placeholder to "Add other poets"