Isolate poems and require authentication for line submission
Add lines route and create endpoint, which adds the line to the database
for the authenticated user, then broadcasts to all connected socket.io
clients. The event name is poem-specific, so lines only appear in the
poems they are intended for.

Although this does require an authenticated user to submit a poem line,
the approach is neither secure nor scalable. Even though the client
subscribes to a poem-specific event, all events are still broadcast to all
clients, and the traffic can be monitored through browser debugging tools.

Socket.io has a concept called "rooms" that might be worth taking a look
at.