Show poets on edit screen, and assign a color to each.
Poets are assigned a color from a palette. client-side logic applies the
poet's color to an indicator in front of each line. Currently there is
unsightly duplication of the DOM between the server and the client, so
next step is to find a way to share DOM rendering logic between the client
and the server.