Render poet colors on the server to eliminate flicker.
There are some tricky bits to this implementation. There is a poet-color partial that contains a function to compute the poet color css class from the poetId. The client needs to provide the partial with a list of poet objects, but it just creates object literals with a poet id property, because that is all that is required by the poetColorClass function.  Would be nice to find a way to avoid having to do this.