Add poems link to logged in user nav
I put the links in a nav element, in a list. I'm not sure about having
the user's name as one of the list items in a nav list, but I'm ok with
it for now. We can move things around once we get a better idea of what
we want the nav to look like.